getting config

+++ src/main/java/capstone/project/recipes/repository/FavoriteRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
+public interface FavoriteRepository extends JpaRepository<Favorite, Integer> {
     List<Favorite> findByUser(User user);
     // Add other query methods as needed
 }
